+++ trixnity-messenger-compose-view/src/commonMain/kotlin/de/connect2x/messenger/compose/view/room/timeline/Timeline.kt
@@ -80,7 +80,6 @@ class TimelineViewImpl : TimelineView {
         val focusManager = LocalFocusManager.current
         val mutableInteractionSource = remember { MutableInteractionSource() }
 
-
         Surface(modifier = Modifier.weight(1.0f, fill = true)) {
             if (timelineElementHolderViewModels.isEmpty()) {
                 Box(Modifier.fillMaxSize()) { CircularProgressIndicator(Modifier.align(Alignment.Center)) }